+++ random/password/variables.tf
@@ -39,12 +39,3 @@ variable "random_min_upper_default" {
   type    = number
   default = 1
 }
-
-variable "std_map" {
-  type = object({
-    name_replace_regex   = string
-    resource_name_prefix = string
-    resource_name_suffix = string
-    tags                 = map(string)
-  })
-}