+++ secret/init_map/locals.tf
@@ -4,19 +4,17 @@ locals {
   }
   l1_map = {
     for k, v in local.l0_map : k => merge(v, {
-      secret_random_init     = v.secret_random_init == null ? var.secret_random_init_default : v.secret_random_init
-      secret_random_init_key = v.secret_random_init_key == null ? var.secret_random_init_key_default : v.secret_random_init_key
-      secret_random_init_map = v.secret_random_init_map == null ? var.secret_random_init_map_default : v.secret_random_init_map
+      has_random_init_value   = lookup(var.secret_random_init_value_map, k, null) != null
+      secret_random_init_key  = v.secret_random_init_key == null ? var.secret_random_init_key_default : v.secret_random_init_key
+      secret_random_init_map  = v.secret_random_init_map == null ? var.secret_random_init_map_default : v.secret_random_init_map
+      secret_random_init_type = v.secret_random_init_type == null ? var.secret_random_init_type_default : v.secret_random_init_type
     })
   }
   l2_map = {
     for k, v in local.l0_map : k => {
-      secret_random_init_map_final = local.l1_map[k].secret_random_init_key == null ? null : merge(
-        local.l1_map[k].secret_random_init_map == null ? {} : local.l1_map[k].secret_random_init_map,
-        {
-          local.l1_map[k].secret_random_init_key = null, # Do not add the secret value here :)
-        }
-      )
+      secret_random_init                  = local.l1_map[k].secret_random_init_type != null
+      secret_random_init_type_is_password = local.l1_map[k].secret_random_init_type == null ? false : local.l1_map[k].secret_random_init_type == "password"
+      secret_random_init_type_is_tls_key  = local.l1_map[k].secret_random_init_type == null ? false : local.l1_map[k].secret_random_init_type == "tls_key"
     }
   }
   lx_map = {