+++ waf/outputs.tf
@@ -1,4 +1,4 @@
 
-output "web_acl_arn" {
-  value = aws_wafv2_web_acl.this_waf_acl.arn
+output "data" {
+  value = local.output_data
 }